+++ requirements.txt
@@ -1,4 +1,1 @@
-dask==2021.5.1
-numpy==1.20.3
 pandas==1.2.4
-
